scripts/requirements-db-test.txt
sqlalchemy>=2.0
asyncpg>=0.29
python-dotenv>=1.0
